Card schema change and minor improvements
Legality now is a String enum. A blank string is used for a legal card, "leader" is used for cards legal as leaders, "banned" is used for cards banned, and "reserved" is used for cards banned due to their status on the reserved list.

Additionally, Human has been removed from the selectable types of a Leader, as it is not a valid Coalition type.